Fix crash if a persons quarantine is checked
When the persons quarantine status is checked a crash is caused if the
person did not have a valid self-test result. This is due to the result
being stored as True/False in case a test was done and None if no test
result is available. This is not correctly handled.